Outlook email from sender nikocoding@outlook.com still not working.
Added admin button, the page doesn't actually link to anything yet but its nice that the button is there now

All we need to do is to get the admin page to actually show some analytics.

We can add the links for the directs of analysis already compiled, but will need to add more including some visuals. Make it meaningful and nice